Add shortcuts import + fix tests
