+++ metabench/algorithmic/metaheuristic/__init__.py
@@ -1,0 +1,1 @@
+from metabench.algorithmic.metaheuristic.abstract import SMetaheuristic
