+++ metabench/algorithmic/metaheuristic/abstract.py
@@ -1,5 +1,5 @@
 """
-File: metaheuristic.py
+File: abstract.py
 Author: Come Bertrand
 Email: bertrand.cosme@gmail.com
 Github: https://github.com/ComeBertrand
