+++ metabench/algorithmic/problem/__init__.py
@@ -1,0 +1,1 @@
+from metabench.algorithmic.problem.abstract import Problem
