+++ metabench/algorithmic/problem/continuous.py
@@ -10,15 +10,17 @@ https://www.sfu.ca/~ssurjano/optimization.html
 
 import numpy as np
 
-from metabench.algorithmic.problem.problem import Problem
-from metabench.common.representation.encoding import RealEncoding, Boundaries
-from metabench.common.objective.objective import Objective
-from metabench.operators.neighborhood.neighborhood import (Neighborhood, move_distance_continuous,
-                                                           ContinuousLogMoveRange)
+from metabench.algorithmic.problem.abstract import Problem
+from metabench.common.representation import RealEncoding, Boundaries
+from metabench.common.fitness import Objective
+from metabench.algorithmic.operators.neighborhood import (NeighborhoodGenerator, move_distance_continuous,
+                                                          ContinuousLogMoveRange)
 
 
 class ContinuousProblem(Problem):
     """Problems that are defined by a continuous function.
+
+    # TODO: Do it in a more abstract way and move it in abstract
 
     Args:
         n_dim (int): Number of dimensions.
@@ -30,19 +32,13 @@ class ContinuousProblem(Problem):
 
     """
     def __init__(self, n_dim, min_vals, max_vals, move_range, known_min):
-        nb_neighbors = n_dim * 100
-        neighborhood = Neighborhood(move_distance_continuous,
-                                    move_range,
-                                    nb_neighbors)
-        boundaries = Boundaries(min_vals,
-                                max_vals,
-                                np.float)
+        nb_neighbors = n_dim * 100 # TODO: shall be an argument of the object
+        neighborhood = NeighborhoodGenerator(move_distance_continuous, move_range, nb_neighbors)
+        boundaries = Boundaries(min_vals, max_vals, np.float)
         encoding = RealEncoding(boundaries)
         objective = Objective(self._eval_func)
-        super().__init__(objective,
-                         encoding,
-                         neighborhood=neighborhood,
-                         known_min=known_min)
+
+        super().__init__(objective, encoding, neighborhood=neighborhood, known_min=known_min)
 
     def _eval_func(self, solution):
         """Actual evaluation of a solution by the continuous function.
