+++ metabench/algorithmic/statistics/__init__.py
@@ -1,0 +1,1 @@
+from metabench.algorithmic.statistics.statistics_recorder import StatisticsRecorder
