+++ metabench/algorithmic/utils/__init__.py
@@ -1,0 +1,1 @@
+from metabench.algorithmic.utils.decorators import not_implemented_for, implemented_for
