+++ metabench/common/fitness/__init__.py
@@ -1,0 +1,2 @@
+from metabench.common.fitness.modif import Modifs
+from metabench.common.fitness.objective import Objective, ObjectiveNoisy
