+++ metabench/common/representation/__init__.py
@@ -1,0 +1,3 @@
+from metabench.common.representation.boundaries import Boundaries
+from metabench.common.representation.encoding import (BinaryEncoding, DiscreteEncoding, RealEncoding,
+                                                      PermutationEncoding)
