+++ metabench/tests/test_decorators.py
@@ -1,7 +1,7 @@
 import pytest
 
-import metabench as mb
 from metabench.tests.fixtures import *
+from metabench.algorithmic.utils import not_implemented_for, implemented_for
 
 
 def test_not_implemented_one_type(fitness_func,
@@ -10,7 +10,7 @@ def test_not_implemented_one_type(fitness_func,
                                   real_solution,
                                   permutation_solution,
                                   one_type_name_permutation):
-    decorator = mb.not_implemented_for(one_type_name_permutation)
+    decorator = not_implemented_for(one_type_name_permutation)
     f = decorator(fitness_func)
     f(binary_solution)
     f(discrete_solution)
@@ -25,7 +25,7 @@ def test_not_implemented_two_type(fitness_func,
                                   real_solution,
                                   permutation_solution,
                                   two_type_names_binary_discrete):
-    decorator = mb.not_implemented_for(*two_type_names_binary_discrete)
+    decorator = not_implemented_for(*two_type_names_binary_discrete)
     f = decorator(fitness_func)
     with pytest.raises(NotImplementedError):
         f(binary_solution)
@@ -41,7 +41,7 @@ def test_implemented_one_type(fitness_func,
                               real_solution,
                               permutation_solution,
                               one_type_name_permutation):
-    decorator = mb.implemented_for(one_type_name_permutation)
+    decorator = implemented_for(one_type_name_permutation)
     f = decorator(fitness_func)
     with pytest.raises(NotImplementedError):
         f(binary_solution)
@@ -58,7 +58,7 @@ def test_implemented_two_type(fitness_func,
                               real_solution,
                               permutation_solution,
                               two_type_names_binary_discrete):
-    decorator = mb.implemented_for(*two_type_names_binary_discrete)
+    decorator = implemented_for(*two_type_names_binary_discrete)
     f = decorator(fitness_func)
     f(binary_solution)
     f(discrete_solution)
