+++ metabench/tests/test_objective.py
@@ -2,10 +2,11 @@ import pytest
 
 import metabench as mb
 from metabench.tests.fixtures import *
+from metabench.common.fitness import *
 
 
 def test_modifs_add_modif(modifs):
-    m = mb.Modifs()
+    m = Modifs()
     for index, val_bef, val_aft in modifs:
         m.add_modif(index, val_bef, val_aft)
 
@@ -15,7 +16,7 @@ def test_modifs_add_modif(modifs):
 
 def test_modifs_add_double_modif(modifs_double):
     modifs, expected = modifs_double
-    m = mb.Modifs()
+    m = Modifs()
     for index, val_bef, val_aft in modifs:
         m.add_modif(index, val_bef, val_aft)
 
@@ -23,7 +24,7 @@ def test_modifs_add_double_modif(modifs_double):
 
 
 def test_modifs_setitem(modifs):
-    m = mb.Modifs()
+    m = Modifs()
     with pytest.raises(NotImplementedError):
         m[0] = ('a', 'b')
 
@@ -32,7 +33,7 @@ def test_objective_no_partial(fitness_func,
                               binary_solution,
                               modifs_as_modifs,
                               modifs_empty):
-    o = mb.Objective(fitness_func)
+    o = Objective(fitness_func)
 
     binary_solution.fitness = None
     assert o._compute_fitness_value(binary_solution, None) == VALUE_RETURNED_FIT
@@ -56,7 +57,7 @@ def test_objective_no_partial_fitness(fitness_func,
                                       binary_solution,
                                       modifs_as_modifs,
                                       modifs_empty):
-    o = mb.Objective(fitness_func)
+    o = Objective(fitness_func)
 
     binary_solution.fitness = VALUE_NOT_RETURNED
     assert o._compute_fitness_value(binary_solution, None) == VALUE_RETURNED_FIT
@@ -81,7 +82,7 @@ def test_objective_partial(fitness_func,
                            binary_solution,
                            modifs_as_modifs,
                            modifs_empty):
-    o = mb.Objective(fitness_func, fitness_partial_func)
+    o = Objective(fitness_func, fitness_partial_func)
 
     binary_solution.fitness = None
     assert o._compute_fitness_value(binary_solution, None) == VALUE_RETURNED_FIT
@@ -106,7 +107,7 @@ def test_objective_partial_fitness(fitness_func,
                                    binary_solution,
                                    modifs_as_modifs,
                                    modifs_empty):
-    o = mb.Objective(fitness_func, fitness_partial_func)
+    o = Objective(fitness_func, fitness_partial_func)
 
     binary_solution.fitness = VALUE_NOT_RETURNED
     assert o._compute_fitness_value(binary_solution, None) == VALUE_RETURNED_FIT
